requirements.txt
Django==4.0
djangorestframework==3.13.0
flake8==4.0.1
psycopg2==2.9.2
drf-yasg==1.20.0